+++ src/pages/onboard/InitProcess.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import InitProcessTemplate from '../../templates/onboarding/InitProcessTemplate';
+import InitProcessTemplate from '../../templates/onboard/InitProcessTemplate';
 
 export default function InitProcess() {
   return (
