+++ src/pages/onboard/SignUp.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignUpTemplate from '../../templates/onboarding/SignUpTemplate';
+import SignUpTemplate from '../../templates/onboard/SignUpTemplate';
 
 export default function SignUp() {
   return (
