+++ src/pages/repair/RepairHome.jsx
@@ -1,0 +1,5 @@
+import RepairTemplate from '../../templates/repair/RepairHomeTemplate';
+
+export default function RepairHome() {
+  return <RepairTemplate />;
+}
