+++ src/templates/onboard/InitProcessTemplate.jsx
