+++ src/templates/repair/RepairHomeTemplate.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function RepairTemplate() {
+export default function RepairHomeTemplate() {
   return (
     <>
       <h1>뚝딱뚜깍</h1>
